requirements.txt
# Core dependencies for ingestion pipeline
google-generativeai>=0.3.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0

# Optional but recommended
requests>=2.31.0